Convert LoginContext to tsx
