+++ src/Components/Home.js
@@ -2,7 +2,7 @@ import { Button, Container, Typography, Box } from "@mui/material";
 import { Link } from "react-router-dom";
 import { useTranslation, Trans } from "react-i18next";
 import { React } from "react";
-import { useLogin } from "./LoginContext";
+import { useLogin } from "./LoginContext.tsx";
 
 const Home = () => {
   const {
